+++ src/main/iptux.cpp
@@ -30,6 +30,7 @@
 #include "iptux/LogSystem.h"
 #include "iptux/SoundSystem.h"
 #include "iptux/support.h"
+#include "iptux/deplib.h"
 
 using namespace std;
 using namespace iptux;
@@ -54,9 +55,37 @@ string getConfigPath() {
         return res2;
 }
 
-int main(int argc, char *argv[])
+static gboolean version = FALSE;
+static gchar* configFilename = nullptr;
+
+static GOptionEntry entries[] =
 {
-  string configPath = getConfigPath();
+  { "version", 'v', 0, G_OPTION_ARG_NONE, &version, "Output version information and exit", NULL },
+  { "config", 'c', 0, G_OPTION_ARG_FILENAME, &configFilename, "Specify config path", "CONFIG_PATH"},
+  { NULL }
+};
+
+int main (int argc, char *argv[]) {
+  setlocale(LC_ALL, "");
+  bindtextdomain(GETTEXT_PACKAGE, __LOCALE_PATH);
+  bind_textdomain_codeset(GETTEXT_PACKAGE, "UTF-8");
+  textdomain(GETTEXT_PACKAGE);
+
+  GError *error = NULL;
+  GOptionContext *context;
+
+  context = g_option_context_new(_("- A software for sharing in LAN"));
+  g_option_context_add_main_entries (context, entries, GETTEXT_PACKAGE);
+  g_option_context_add_group (context, gtk_get_option_group (TRUE));
+  if (!g_option_context_parse (context, &argc, &argv, &error)) {
+    g_print(_("option parsing failed: %s\n"), error->message);
+    exit (1);
+  }
+  if(version) {
+    printf("iptux: " VERSION "\n");
+    exit(0);
+  }
+  string configPath = configFilename ? configFilename : getConfigPath();
   IptuxConfig config(configPath);
   ProgramData progdt(config);
   MainWindow mwin(config, progdt);
@@ -73,13 +102,6 @@ int main(int argc, char *argv[])
 
 
   mwin.SetStatusIcon(&sicon);
-
-        setlocale(LC_ALL, "");
-        bindtextdomain(GETTEXT_PACKAGE, __LOCALE_PATH);
-        bind_textdomain_codeset(GETTEXT_PACKAGE, "UTF-8");
-        textdomain(GETTEXT_PACKAGE);
-
-        analysis_parameter(argc, argv);
 
         gdk_threads_init();
         gdk_threads_enter();